requirements.txt
airtable-python-wrapper==0.15.2
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-slugify==5.0.2
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
yapf==0.31.0
